[APDV-111] Moved EndpointMapper
