+++ backend/src/main/java/pl/edu/agh/apdvbackend/use_cases/endpoint/SaveNewEndpointImpl.java
@@ -2,7 +2,7 @@ package pl.edu.agh.apdvbackend.use_cases.endpoint;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
-import pl.edu.agh.apdvbackend.mappers.EndpointMapper;
+import pl.edu.agh.apdvbackend.mappers.endpoint.EndpointMapper;
 import pl.edu.agh.apdvbackend.models.body_models.endpoint.AddEndpointRequestBody;
 import pl.edu.agh.apdvbackend.models.database.Endpoint;
 import pl.edu.agh.apdvbackend.repositories.EndpointRepository;
